Add a CRC32C method to platform.
This is not a drop-in replacement for existing CRC functions
in Couchbase. SSE4.2 instructions use a incompatible polynomial
so code must be aware of that if they have any persisted
checksums to examine.

The implementation is based upon Mark Adler's version
(acknowledged in the source).

Tuned, validated and made compatible for all Couchbase platforms.

When available SSE4.2 hardware assistance is used.

Change-Id: I3fc881c7712de696a6714a59cb350b991c458032
Reviewed-on: http://review.couchbase.org/53336
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>

Conflicts:
	CMakeLists.txt